udpates for gear version

+++ sage_flywheel_fixbids.py
@@ -16,25 +16,29 @@ inner_folder = "%s/ses-EichSAGE" % subject_folder
 fmap_folder = "%s/fmap" % inner_folder
 code_dir = os.path.dirname(os.path.realpath(__file__))
 
-# if not os.path.exists(fmap_folder):
-#     os.mkdir(fmap_folder)
+if not os.path.exists(fmap_folder):
+    os.mkdir(fmap_folder)
 
 # Fix capitalization of anatomicals
 src_file = "%s/anat/sub-%s_ses-EichSAGE_t1w.json" % (inner_folder,subject_code)
 dest_file = "%s/anat/sub-%s_ses-EichSAGE_T1w.json" % (inner_folder,subject_code)
-shutil.move(src_file,dest_file)
+if (os.path.exists(src_file)):
+    shutil.move(src_file,dest_file)
 
 src_file = "%s/anat/sub-%s_ses-EichSAGE_t1w.nii.gz" % (inner_folder,subject_code)
 dest_file = "%s/anat/sub-%s_ses-EichSAGE_T1w.nii.gz" % (inner_folder,subject_code)
-shutil.move(src_file,dest_file)
+if (os.path.exists(src_file)):
+    shutil.move(src_file,dest_file)
 
 src_file = "%s/anat/sub-%s_ses-EichSAGE_t2w.json" % (inner_folder,subject_code)
 dest_file = "%s/anat/sub-%s_ses-EichSAGE_T2w.json" % (inner_folder,subject_code)
-shutil.move(src_file,dest_file)
+if (os.path.exists(src_file)):
+    shutil.move(src_file,dest_file)
 
 src_file = "%s/anat/sub-%s_ses-EichSAGE_t2w.nii.gz" % (inner_folder,subject_code)
 dest_file = "%s/anat/sub-%s_ses-EichSAGE_T2w.nii.gz" % (inner_folder,subject_code)
-shutil.move(src_file,dest_file)
+if (os.path.exists(src_file)):
+    shutil.move(src_file,dest_file)
 
 #Copy files to be used as field maps
 src_file = "%s/func/sub-%s_ses-EichSAGE_task-hiAP_run-01_bold.nii.gz" % (inner_folder,subject_code)
